Changed name of player Prefab, added the shotgun as a weapon, made enemy spawner a prefab
Player spawn prefab name -> PlayerSpawner
Shotgun model and stats are now a prefab that can be used in game
enemy spawner is now a prefab

+++ Assets/Scripts/GunStats.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [CreateAssetMenu]
 public class GunStats : ScriptableObject
 {
-    public int shootDamage;
+    public float shootDamage;
     public float shootRate;
     public int shootDistance;
     public GameObject gunModel;
